refactor: 하나의 framework 패키지로 묶기
