+++ src/main/java/baseball/Application.java
@@ -1,8 +1,12 @@
 package baseball;
 
+import game.Game;
+import game.GameExecutor;
+
 public class Application {
+
     public static void main(String[] args) {
-        BaseballGame game = new BaseballGame(new MissionUtilsGameRandom(), new ConsoleInput(), new ConsoleLogger());
+        Game game = new BaseballGame(new MissionUtilsGameRandom(), new ConsoleInput(), new ConsoleLogger());
         GameExecutor gameExecutor = new GameExecutor(game);
         gameExecutor.run();
     }
