+++ src/main/java/baseball/ConsoleInput.java
@@ -1,6 +1,7 @@
 package baseball;
 
 import camp.nextstep.edu.missionutils.Console;
+import game.GameInputInterface;
 
 public class ConsoleInput implements GameInputInterface {
     @Override
