+++ src/main/java/baseball/ConsoleLogger.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import game.GameOutputInterface;
+
 public class ConsoleLogger implements GameOutputInterface {
     @Override
     public void print(String s) {
