+++ src/main/java/game/Game.java
@@ -1,4 +1,4 @@
-package baseball;
+package game;
 
 public interface Game {
     Game createNewGame();
