+++ src/main/java/game/GameInputInterface.java
@@ -1,4 +1,4 @@
-package baseball;
+package game;
 
 public interface GameInputInterface {
     String getInput();
