+++ src/main/java/game/GameOutputInterface.java
@@ -1,4 +1,4 @@
-package baseball;
+package game;
 
 public interface GameOutputInterface {
     void print(String s);
