+++ src/main/java/game/GameRandom.java
@@ -1,4 +1,4 @@
-package baseball;
+package game;
 
 public interface GameRandom {
     int[] sampleDistinctNumbers(int begin, int end, int length);
